Fix stream name sorting order
